Use jest projects instead of custom jest wrapper
