+++ package.json
@@ -103,9 +103,6 @@
   "devEngines": {
     "node": "16.x || 18.x || 20.x || 21.x"
   },
-  "jest": {
-    "testRegex": "/scripts/jest/dont-run-jest-directly\\.js$"
-  },
   "scripts": {
     "build": "node ./scripts/rollup/build-all-release-channels.js",
     "build-for-devtools": "cross-env RELEASE_CHANNEL=experimental yarn build react/index,react/jsx,react-dom/index,react-dom/unstable_testing,react-dom/test-utils,react-is,react-debug-tools,scheduler,react-test-renderer,react-refresh,react-art --type=NODE",
@@ -116,10 +113,7 @@
     "lint-build": "node ./scripts/rollup/validate/index.js",
     "extract-errors": "node scripts/error-codes/extract-errors.js",
     "postinstall": "node node_modules/fbjs-scripts/node/check-dev-engines.js package.json && node ./scripts/flow/createFlowConfigs.js",
-    "test": "node ./scripts/jest/jest-cli.js",
-    "test-stable": "node ./scripts/jest/jest-cli.js --release-channel=stable",
-    "test-www": "node ./scripts/jest/jest-cli.js --release-channel=www-modern",
-    "test-classic": "node ./scripts/jest/jest-cli.js --release-channel=www-classic",
+    "test": "jest",
     "test-build-devtools": "node ./scripts/jest/jest-cli.js --build --project devtools --release-channel=experimental",
     "test-dom-fixture": "cd fixtures/dom && yarn && yarn predev && yarn test",
     "flow": "node ./scripts/tasks/flow.js",
