--- scripts/jest/config.base.js
@@ -1,32 +1,0 @@
-'use strict';
-
-module.exports = {
-  globalSetup: require.resolve('./setupGlobal.js'),
-  modulePathIgnorePatterns: [
-    '<rootDir>/scripts/rollup/shims/',
-    '<rootDir>/scripts/bench/',
-  ],
-  transform: {
-    '.*': require.resolve('./preprocessor.js'),
-  },
-  prettierPath: require.resolve('prettier-2'),
-  setupFiles: [require.resolve('./setupEnvironment.js')],
-  setupFilesAfterEnv: [require.resolve('./setupTests.js')],
-  // Only include files directly in __tests__, not in nested folders.
-  testRegex: '/__tests__/[^/]*(\\.js|\\.coffee|[^d]\\.ts)$',
-  moduleFileExtensions: ['js', 'json', 'node', 'coffee', 'ts'],
-  rootDir: process.cwd(),
-  roots: ['<rootDir>/packages', '<rootDir>/scripts'],
-  collectCoverageFrom: ['packages/**/*.js'],
-  fakeTimers: {
-    enableGlobally: true,
-    legacyFakeTimers: true,
-  },
-  snapshotSerializers: [require.resolve('jest-snapshot-serializer-raw')],
-
-  testSequencer: require.resolve('./jestSequencer'),
-
-  testEnvironment: 'jsdom',
-
-  testRunner: 'jest-circus/runner',
-};
