--- scripts/jest/config.source-persistent.js
@@ -1,21 +1,0 @@
-'use strict';
-
-const baseConfig = require('./config.base');
-
-module.exports = Object.assign({}, baseConfig, {
-  modulePathIgnorePatterns: [
-    ...baseConfig.modulePathIgnorePatterns,
-    'packages/react-devtools-extensions',
-    'packages/react-devtools-shared',
-    'ReactIncrementalPerf',
-    'ReactIncrementalUpdatesMinimalism',
-    'ReactIncrementalTriangle',
-    'ReactIncrementalReflection',
-    'forwardRef',
-  ],
-  setupFiles: [
-    ...baseConfig.setupFiles,
-    require.resolve('./setupTests.persistent.js'),
-    require.resolve('./setupHostConfigs.js'),
-  ],
-});
