--- scripts/jest/config.source.js
@@ -1,15 +1,0 @@
-'use strict';
-
-const baseConfig = require('./config.base');
-
-module.exports = Object.assign({}, baseConfig, {
-  modulePathIgnorePatterns: [
-    ...baseConfig.modulePathIgnorePatterns,
-    'packages/react-devtools-extensions',
-    'packages/react-devtools-shared',
-  ],
-  setupFiles: [
-    ...baseConfig.setupFiles,
-    require.resolve('./setupHostConfigs.js'),
-  ],
-});
