--- scripts/jest/dont-run-jest-directly.js
@@ -1,3 +1,0 @@
-'use strict';
-
-throw new Error("Don't run `jest` directly. Run `yarn test` instead.");
