+++ scripts/jest/setupEnvDevelopment.js
@@ -1,0 +1,3 @@
+'use strict';
+
+process.env.NODE_ENV = 'development';
