+++ scripts/jest/setupEnvVariant.js
@@ -1,0 +1,3 @@
+'use strict';
+
+process.env.VARIANT = 'true';
